+++ src/Cms/CmsService.php
@@ -50,6 +50,12 @@ final class CmsService extends AbstractService
 
     private function convertArrayToMenuItem(array $menuItem): MenuItem
     {
-        return new MenuItem(to_string($menuItem['name']), to_int($menuItem['position']), new Uri(to_string($menuItem['url'])), array_map([$this, 'convertArrayToMenuItem'], $menuItem['children']));
+        return new MenuItem(
+            to_string($menuItem['name']),
+            to_int($menuItem['position']),
+            new Uri(to_string($menuItem['url'])),
+            (bool) $menuItem['targetBlank'],
+            array_map([$this, 'convertArrayToMenuItem'], $menuItem['children'])
+        );
     }
 }